+++ src/Route.jsx
@@ -10,24 +10,7 @@ import {
 } from 'Pages';
 import { SideBar } from 'Components';
 
-import { AuthContext } from './App';
 import Styled from 'Styled/Global.styled';
-
-const Loading = () => {
-  return <div>로딩중!!!</div>;
-};
-
-const OAuthCheckRoute = ({ children }) => {
-  const auth = useContext(AuthContext);
-  console.log('auth : ', auth);
-  if (auth.isLoading) {
-    return <Loading />;
-  } else {
-    if (auth.state !== 401) return children;
-    else return <Navigate to="/login" />;
-    // return children;
-  }
-};
 
 const MainRoute = () => {
   return (
@@ -37,44 +20,13 @@ const MainRoute = () => {
       </Styled.CusTab>
       <Styled.Body>
         <Routes>
-          <Route
-            path="/"
-            element={
-              <OAuthCheckRoute>
-                <MainPage />
-              </OAuthCheckRoute>
-            }
-          />
+          <Route path="/" element={<MainPage />} />
 
-          <Route
-            path="/todo"
-            element={
-              <OAuthCheckRoute>
-                <TodoPage />
-              </OAuthCheckRoute>
-            }
-          />
+          <Route path="/todo" element={<TodoPage />} />
 
-          <Route
-            path="/mine"
-            element={
-              <OAuthCheckRoute>
-                <MinePage />
-              </OAuthCheckRoute>
-            }
-          />
+          <Route path="/mine" element={<MinePage />} />
 
-          <Route
-            path="/admin"
-            element={
-              <OAuthCheckRoute>
-                <AdminPage />
-              </OAuthCheckRoute>
-            }
-          />
-
-          <Route path="/login" element={<LoginPage />} />
-          <Route path="/oauth/kakao/callback" element={<OAuthPage />} />
+          <Route path="/admin" element={<AdminPage />} />
         </Routes>
       </Styled.Body>
     </>
